Refactor rename .js to .jsx and remove unused imports
