+++ src/components/buttons/AddButton.jsx
@@ -1,9 +1,7 @@
 const AddButton = ({ call }) => {
-
   return (
     <button className="btn btn-primary" onClick={call}>+</button>
   );
 }
 
 export default AddButton;
-
