+++ src/components/buttons/BackButton.jsx
@@ -1,10 +1,8 @@
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { BiArrowBack } from 'react-icons/bi';
 
 const BackButton = () => {
-
   const navigate = useNavigate();
-
   const goBack = () => navigate(-1);
 
   return (
