+++ src/components/buttons/DeleteButton.jsx
@@ -1,7 +1,6 @@
 import { BsTrashFill } from 'react-icons/bs';
 
 const DeleteButton = ({ setShowDeleteWarning }) => {
-
   return (
     <button className="btn" onClick={() => setShowDeleteWarning(true)}>
       <BsTrashFill/>
@@ -10,4 +9,3 @@ const DeleteButton = ({ setShowDeleteWarning }) => {
 }
 
 export default DeleteButton;
-
