+++ src/components/buttons/EditButton.jsx
@@ -1,7 +1,6 @@
 import { BsPencilFill } from 'react-icons/bs';
 
 const EditButton = ({call}) => {
-
   return (
     <button className="btn" onClick={call}>
       <BsPencilFill/>
