+++ src/components/layout/TopNav.jsx
@@ -1,11 +1,11 @@
 import { Navbar, Container } from "react-bootstrap";
 import { useLocation } from "react-router-dom";
 import { pages } from '../../services/enums';
-import Visibility from '../../components/containers/Visibility';
-import BackButton from "../../components/buttons/BackButton";
-import InstallButton from "../../components/buttons/InstallButton";
+import Visibility from '../containers/Visibility';
+import BackButton from "../buttons/BackButton";
+import InstallButton from "../buttons/InstallButton";
 import { parentPageOf } from '../../services/parentPage';
-import Settings from '../../components/layout/Settings';
+import Settings from './Settings';
 
 const TopNav = ({ setShowAboutPage, setShowResetWarning }) => {
   const parentPage = parentPageOf(useLocation().pathname.toString().toLowerCase());
