--- src/components/misc/FreqList.js
@@ -1,18 +1,0 @@
-import { stickFrequencies } from '../../services/enums';
-
-const FreqList = ({ stickable }) => {
-
-  return (
-    <>
-    {stickFrequencies.map((item, i) => {
-      return (
-        <label className="skillFreq" freq={i} key={i}>
-          <input type="radio" value={i} name="stickFrequency" checked={(stickable.stickFrequency === i)} readOnly={true} /> {item}
-        </label>
-      );
-    })}
-    </>
-  );
-}
-
-export default FreqList;
